# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UptimeKit API
  description: REST API for UptimeKit HTTP Monitors and Authentication
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Monitor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        expected_status:
          type: integer
        check_interval:
          type: integer
        status:
          type: string
          enum: [up, down, degraded, paused]
        http_method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS]
        is_deleted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - team_id
        - name
        - url
        - expected_status
        - check_interval
        - status
        - http_method
        - is_deleted
        - created_at
        - updated_at
    CreateMonitorRequest:
      type: object
      required:
        - url
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        expected_status:
          type: integer
          default: 200
        check_interval:
          type: integer
          default: 60
        http_method:
          type: string
          default: GET
    UpdateMonitorRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        expected_status:
          type: integer
        check_interval:
          type: integer
        http_method:
          type: string
        status:
          type: string
          enum: [paused, up]
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
paths:
  /auth/login:
    post:
      summary: Login to get access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Refresh token cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /auth/refresh:
    post:
      summary: Refresh access token
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
  /auth/logout:
    post:
      summary: Logout and clear refresh token
      responses:
        '200':
          description: Successfully logged out
  /monitors:
    get:
      summary: List all monitors
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
            format: uuid
          description: UUIDv7 cursor for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: A list of monitors
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Monitor'
                  nextCursor:
                    type: string
                    format: uuid
                    nullable: true
        '401':
          description: Unauthorized
    post:
      summary: Create a new monitor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMonitorRequest'
      responses:
        '201':
          description: Monitor created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Monitor'
        '400':
          description: Validation error or SSRF rejection
        '401':
          description: Unauthorized
  /monitors/{id}:
    get:
      summary: Get a monitor by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Monitor details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Monitor'
        '401':
          description: Unauthorized
        '404':
          description: Monitor not found
    patch:
      summary: Update a monitor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMonitorRequest'
      responses:
        '200':
          description: Monitor updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Monitor'
        '400':
          description: Validation error or SSRF rejection
        '401':
          description: Unauthorized
        '404':
          description: Monitor not found
    delete:
      summary: Soft delete a monitor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Monitor deleted
        '401':
          description: Unauthorized
        '404':
          description: Monitor not found